+++ README.md
@@ -2,4 +2,5 @@
 
 ## Getting Started
 1. Create a `.env` file from the `.env.example` file. And fill in the information
-2. Simply run `npm start`
+2. If you modify the js/index.ts, please use "tsc" to compile the project at root level (README.md level)
+2. Simply run `npm start`at root level (README.md level)
