+++ package.json
@@ -1,11 +1,11 @@
 {
   "name": "camera",
-  "version": "1.0.0",
+  "version": "0.1.0",
   "description": "",
   "main": "index.js",
   "scripts": {
     "start": "node index.js",
-    "dev": "webpack --watch",
+    "dev": "tsc --watch",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
@@ -19,10 +19,9 @@
     "project-oxford": "^1.7.0"
   },
   "devDependencies": {
-    "babel-core": "^6.23.1",
-    "babel-loader": "^6.3.2",
-    "babel-preset-es2015": "^6.22.0",
-    "query-string": "^4.3.2",
-    "webpack": "^2.2.1"
+    "@types/node": "^7.0.5",
+    "source-map-loader": "^0.1.5",
+    "ts-loader": "^0.8.2",
+    "typescript": "^2.0.3"
   }
 }
