+++ public/index.html
@@ -13,6 +13,6 @@
     <label for="personId">Person ID</label>
     <input id="personId" />
     <button id="trainButton">Train!</button>
-    <script src='build/bundle.js'></script>
+    <script src='build/index.js'></script>
 </body>
 </html>
