+++ tsconfig.json
@@ -1,0 +1,13 @@
+{
+    "compilerOptions": {
+        "module": "commonjs",
+        "target": "es6",
+        "noImplicitAny": false,
+        "sourceMap": false,
+        "allowJs": true,
+        "outDir": "public/build"
+    },
+    "files": [
+        "./public/js/index.ts"
+    ]
+}
