--- webpack.config.js
@@ -1,21 +1,0 @@
-var path = require('path')
-
-module.exports = {
-  entry: './public/js/index.js',
-  module: {
-    loaders: [
-      {
-        test: /\.js$/,
-        exclude: /node_modules/,
-        loader: 'babel-loader',
-        query: {
-          presets: ['es2015']
-        }
-      }
-    ]
-  },
-  output: {
-    filename: 'bundle.js',
-    path: path.resolve(__dirname, 'public/build')
-  }
-}
